pyproject.toml
[tool.poetry]
name = "wyc"
version = "0.1.0"
description = ""
authors = ["abdulazizmirjalilov <abdulaziz.mirjalilov@ventionteams.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
pydantic = "^2.11.5"
gunicorn = "^23.0.0"
uvicorn = "^0.34.3"
asyncpg = "^0.30.0"
alembic = "^1.16.1"
sqlalchemy = "^2.0.41"
pydantic-settings = "^2.9.1"
openai = "^1.86.0"
livekit-agents = "^1.1.1"
livekit-plugins-openai = "^1.1.1"
livekit-plugins-silero = "^1.1.1"
python-dotenv = "^1.1.0"
livekit-api = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
